+++ src/main/java/cart/controller/CartApiController.java
@@ -1,8 +1,6 @@
 package cart.controller;
 
 import cart.dto.auth.AuthInfo;
-import cart.dto.request.RequestCreateProductDto;
-import cart.dto.request.RequestUpdateProductDto;
 import cart.dto.response.ResponseProductDto;
 import cart.infrastructure.BasicAuthorizationExtractor;
 import cart.service.CartService;
@@ -12,8 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
-import java.net.URI;
 import java.util.List;
 
 @RestController
@@ -29,24 +25,6 @@ public class CartApiController {
         this.memberService = memberService;
     }
 
-    @PostMapping("/product")
-    public ResponseEntity<Void> createProduct(@RequestBody @Valid final RequestCreateProductDto requestCreateProductDto) {
-        cartService.insert(requestCreateProductDto);
-        return ResponseEntity.created(URI.create("/product")).build();
-    }
-
-    @PutMapping("/product/{id}")
-    public ResponseEntity<Void> updateProduct(@PathVariable final Long id, @RequestBody @Valid final RequestUpdateProductDto requestUpdateProductDto) {
-        cartService.update(id, requestUpdateProductDto);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/product/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable final Long id) {
-        cartService.delete(id);
-        return ResponseEntity.ok().build();
-    }
-
     @GetMapping("/carts")
     public ResponseEntity<List<ResponseProductDto>> getProducts(final HttpServletRequest request) {
         final AuthInfo authIfo = basicAuthorizationExtractor.extract(request);
@@ -54,7 +32,7 @@ public class CartApiController {
         return ResponseEntity.ok().body(cartProductsByMember);
     }
 
-    @PutMapping("/products/{id}")
+    @PutMapping("/cart/{id}")
     public ResponseEntity<Void> addProductToCart(@PathVariable final Long id, final HttpServletRequest request) {
         final AuthInfo authInfo = basicAuthorizationExtractor.extract(request);
         final Long memberId = memberService.findIdByAuthInfo(authInfo);
@@ -62,7 +40,7 @@ public class CartApiController {
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/products/{id}")
+    @DeleteMapping("/cart/{id}")
     public ResponseEntity<Void> deleteProductFromCart(@PathVariable final Long id, final HttpServletRequest request) {
         final AuthInfo authInfo = basicAuthorizationExtractor.extract(request);
         final Long memberId = memberService.findIdByAuthInfo(authInfo);
