+++ src/main/resources/templates/index.html
@@ -14,7 +14,7 @@
             <li><a href="/">상품목록</a></li>
             <li><a href="/cart">장바구니</a></li>
             <li><a href="/settings">설정</a></li>
-            <li class="nav-admin"><a href="/">관리자</a></li>
+            <li class="nav-admin"><a href="/admin">관리자</a></li>
         </ul>
     </nav>
 </header>
